Improve exit status of cynagora-admin
When used command line, the tool cynagora-admin
was terminating with status 0 even when the server
wasn't up and running.

This changes corrctly fix the issue by recording
the status of the last cynagora command and by
treating with it in command line mode.

This change is required to implement a basic ptest
and to submit cynagora for Itchy Icefish.

Bug-AGL: 2844

Also fix a naming in the cache stressing test.

Change-Id: I39ee784bbe8ce7b792e96ef14578baa6f80ef6ac
Signed-off-by: José Bollo <jose.bollo@iot.bzh>